pyproject.toml
[project]
name = "BusArrivalBot"
version = "0.1.0"
description = "Telegram bot for checking bus arrival times in Singapore"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "python-decouple>=3.8",
    "python-telegram-bot>=22.3",
    "requests>=2.32.4",
    "scipy>=1.16.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.7",
]
